correcting many bugs and managing better values

+++ src/prv2paje/headers/pajepending.h
@@ -17,7 +17,7 @@ namespace prv2paje{
     public:
         PajePending();
         void addPajePendingEvent(PajePendingEvent *pajePendingEvent);
-        void pushPendingEvents(long int timestamp);
+        void pushPendingEvents(double timestamp);
     private:
         list<PajePendingEvent*> pajePendingEvents;
     };
